Organize tests and enable coverage
