+++ tests/unit/test_dedup.py
@@ -1,10 +1,14 @@
 import os
 import sys
 import pandas as pd
-import numpy as np
+import pytest
 
-sys.path.insert(0, os.path.abspath(os.path.join(os.path.dirname(__file__), "..", "src")))
-import app as appmod
+pytestmark = pytest.mark.unit
+
+sys.path.insert(
+    0, os.path.abspath(os.path.join(os.path.dirname(__file__), "..", "..", "src"))
+)
+import app as appmod  # noqa: E402
 
 
 def test_select_latest_priority():
