+++ tests/unit/test_periods.py
@@ -1,11 +1,13 @@
 import os
 import sys
-import types
-import pandas as pd
 import pytest
 
-sys.path.insert(0, os.path.abspath(os.path.join(os.path.dirname(__file__), "..", "src")))
-import app as appmod
+pytestmark = pytest.mark.unit
+
+sys.path.insert(
+    0, os.path.abspath(os.path.join(os.path.dirname(__file__), "..", "..", "src"))
+)
+import app as appmod  # noqa: E402
 
 
 def test_periods_years_annual():
