+++ tests/unit/test_save_naming.py
@@ -1,9 +1,14 @@
 import os
 import sys
 import pandas as pd
+import pytest
 
-sys.path.insert(0, os.path.abspath(os.path.join(os.path.dirname(__file__), "..", "src")))
-import app as appmod
+pytestmark = pytest.mark.unit
+
+sys.path.insert(
+    0, os.path.abspath(os.path.join(os.path.dirname(__file__), "..", "..", "src"))
+)
+import app as appmod  # noqa: E402
 
 
 def test_save_naming(tmp_path):
